requirements.txt
asgiref==3.7.2
Django==5.0.1
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
